Move DataAssociator base class methods to GlobalNearestNeighbour
These methods are only used for GNN, and don't seem relevant to other
data associators, as they use brute force unoptimised approach (which in
part is what having the GNN is there to demonstrate).
