+++ stonesoup/dataassociator/neighbour.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
+import itertools
+
 import numpy as np
 
 from .base import DataAssociator
 from ._assignment import assign2D
 from ..base import Property
 from ..hypothesiser import Hypothesiser
-from ..types.hypothesis import SingleHypothesis, SingleProbabilityHypothesis
+from ..types.hypothesis import SingleHypothesis, SingleProbabilityHypothesis, JointHypothesis
 
 
 class NearestNeighbour(DataAssociator):
@@ -107,6 +109,65 @@ class GlobalNearestNeighbour(DataAssociator):
         associations = max(joint_hypotheses)
 
         return associations
+
+    @staticmethod
+    def isvalid(joint_hypothesis):
+        """Determine whether a joint_hypothesis is valid.
+
+        Check the set of hypotheses that define a joint hypothesis to ensure a
+        single detection is not associated to more than one track.
+
+        Parameters
+        ----------
+        joint_hypothesis : :class:`JointHypothesis`
+            A set of hypotheses linking each prediction to a single detection
+
+        Returns
+        -------
+        bool
+            Whether joint_hypothesis is a valid set of hypotheses
+        """
+
+        number_hypotheses = len(joint_hypothesis)
+        unique_hypotheses = len(
+            {hyp.measurement for hyp in joint_hypothesis if hyp})
+        number_null_hypotheses = sum(not hyp for hyp in joint_hypothesis)
+
+        # joint_hypothesis is invalid if one detection is assigned to more than
+        # one prediction. Multiple missed detections are valid.
+        if unique_hypotheses + number_null_hypotheses == number_hypotheses:
+            return True
+        else:
+            return False
+
+    @classmethod
+    def enumerate_joint_hypotheses(cls, hypotheses):
+        """Enumerate the possible joint hypotheses.
+
+        Create a list of all possible joint hypotheses from the individual
+        hypotheses and determine whether each is valid.
+
+        Parameters
+        ----------
+        hypotheses : list of :class:`Hypothesis`
+            A list of all hypotheses linking predictions to detections,
+            including missed detections
+
+        Returns
+        -------
+        joint_hypotheses : list of :class:`JointHypothesis`
+            A list of all valid joint hypotheses with a score on each
+        """
+
+        # Create a list of dictionaries of valid track-hypothesis pairs
+        joint_hypotheses = [
+            JointHypothesis({
+                track: hypothesis
+                for track, hypothesis in zip(hypotheses, joint_hypothesis)})
+            for joint_hypothesis in itertools.product(*hypotheses.values())
+            if cls.isvalid(joint_hypothesis)]
+
+        return joint_hypotheses
 
 
 class GNNWith2DAssignment(DataAssociator):
